backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
alpaca-py==0.26.0
pandas==2.2.3
pandas-ta==0.3.14b
anthropic==0.34.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
yfinance==0.2.44
stripe==10.12.0
praw==7.7.1
tweepy==4.14.0
